Quote left square brackets in reminder text
If someone puts a bracketed expression in the summary or location of a
calendar entry, then remind will interpret it as a substitution
expression, so quote it. As per the remind man page, the only way to
include a left bracket in a remind string is itself as a substituion
expression ["["], so that's what we do. Note that as long as the left
bracket is quoted, the right one doesn't need to be, since it is only
special when preceded by a left bracket.