End times should be outside of %"...%"
Although the remind documentation claims that remind doesn't actually
use `DURATION` for anything, that's not entirely accurate. It turns
out that both `rem -c` and `rem2ps` display reminder end times when
both `AT` and `DURATION` are both specified. Therefore, it's not
necessary to put the end time explicitly inside the %"...%" calendar
markers.